requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
psycopg2-binary==2.9.9
python-multipart==0.0.6
httpx==0.25.1
asyncio==3.4.3
sse-starlette==1.6.5
python-dotenv==1.0.0